[CALCITE-3782] Moving Bitwise utils to utility functios
